修复了ring mask错误，它超过hough参数规定的范围
